fix(#134): move outdoor-outlook-model to shared presenter layer
Move cross-cutting outdoor outlook logic from email presenter to shared
presenter layer to eliminate presenter-to-presenter coupling.

Changes:
- Move outdoor-comfort.ts and outdoor-outlook-model.ts to src/presenters/shared/
- Update imports in site/sections/hourly-outlook.ts to use shared path
- Update imports in email/render-outdoor-outlook.ts to use shared path
- Update imports in email/next-day.ts to re-export from shared
- Convert email/outdoor-comfort.ts and email/outdoor-outlook-model.ts to
  backwards-compatible re-export modules with deprecation notices
- Update README files to document the new structure
- Move corresponding test files to shared/ folder

This resolves the architectural violation where the site presenter was
importing from the email presenter's internal modules.

Closes #134
